requirements.txt
openai==0.28
python-dotenv
flask
flask-cors